chore: move ignoreStatusUpdatesAndDeletion to own file
